Move out MarketSection to component Also use container pattern for pool header
